Start tests and rename create asset register user case folder
