+++ AcceptanceTest/HomesEngland/UseCases/CalculateAssetAggregates/CalculateAssetAggregatesUseCaseAcceptanceTests.cs
@@ -6,11 +6,11 @@ using System.Transactions;
 using FluentAssertions;
 using HomesEngland.Gateway;
 using HomesEngland.Gateway.Migrations;
-using HomesEngland.UseCase.BulkCreateAsset;
 using HomesEngland.UseCase.CalculateAssetAggregates;
 using HomesEngland.UseCase.CalculateAssetAggregates.Models;
 using HomesEngland.UseCase.CreateAsset;
 using HomesEngland.UseCase.CreateAsset.Models;
+using HomesEngland.UseCase.CreateAssetRegisterVersion;
 using HomesEngland.UseCase.GetAsset.Models;
 using Main;
 using Microsoft.EntityFrameworkCore;
