+++ HomesEngland.Gateway.Test/AssetAggregatorGatewayTests.cs
@@ -10,8 +10,8 @@ using HomesEngland.Gateway.AssetRegisterVersions;
 using HomesEngland.Gateway.Assets;
 using HomesEngland.Gateway.Migrations;
 using HomesEngland.Gateway.Sql;
-using HomesEngland.UseCase.BulkCreateAsset.Models;
 using HomesEngland.UseCase.CalculateAssetAggregates.Models;
+using HomesEngland.UseCase.CreateAssetRegisterVersion.Models;
 using Microsoft.EntityFrameworkCore;
 using NUnit.Framework;
 using TestHelper;
