+++ HomesEngland/Gateway/AssetRegisterVersions/IAssetRegisterVersionCreator.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
-using HomesEngland.UseCase.BulkCreateAsset.Models;
+using HomesEngland.UseCase.CreateAssetRegisterVersion.Models;
 
 namespace HomesEngland.Gateway.AssetRegisterVersions
 {
