--- HomesEngland/UseCase/BulkCreateAsset/ICreateAssetRegisterVersionUseCase.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-using HomesEngland.Boundary.UseCase;
-using HomesEngland.UseCase.CreateAsset.Models;
-
-namespace HomesEngland.UseCase.BulkCreateAsset
-{
-    public interface ICreateAssetRegisterVersionUseCase : IAsyncUseCaseTask<IList<CreateAssetRequest>, IList<CreateAssetResponse>>
-    {
-    }
-}
