+++ HomesEngland/UseCase/CreateAssetRegisterVersion/Models/IAssetRegisterVersion.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using HomesEngland.Domain;
+
+namespace HomesEngland.UseCase.CreateAssetRegisterVersion.Models
+{
+    public interface IAssetRegisterVersion : IDatabaseEntity<int>
+    {
+        IList<IAsset> Assets { get; set; }
+    }
+}
