+++ HomesEngland/UseCase/SaveUploadedAssetRegisterFile/Models/SaveAssetRegisterFileRequest.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace HomesEnglandTest.UseCase.SaveUploadedAssetRegisterFile.Models
+{
+    public class SaveAssetRegisterFileRequest
+    {
+        public string FileName { get; set; }
+        public string FilePath { get; set; }
+    }
+}
