+++ HomesEngland/UseCase/SaveUploadedAssetRegisterFile/Models/SaveAssetRegisterFileResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace HomesEnglandTest.UseCase.SaveUploadedAssetRegisterFile.Models
+{
+    public class SaveAssetRegisterFileResponse
+    {
+
+    }
+}
