+++ HomesEnglandTest/UseCase/CreateAsset/BulkCreateAssetTests.cs
@@ -6,9 +6,9 @@ using FluentAssertions;
 using HomesEngland.Domain;
 using HomesEngland.Exception;
 using HomesEngland.Gateway.AssetRegisterVersions;
-using HomesEngland.UseCase.BulkCreateAsset;
-using HomesEngland.UseCase.BulkCreateAsset.Models;
 using HomesEngland.UseCase.CreateAsset.Models;
+using HomesEngland.UseCase.CreateAssetRegisterVersion;
+using HomesEngland.UseCase.CreateAssetRegisterVersion.Models;
 using Moq;
 using NUnit.Framework;
 using TestHelper;
