+++ HomesEnglandTest/UseCase/SaveUploadedAssetRegisterFile/SaveUploadedAssetRegisterFileUseCaseTests.cs
@@ -1,0 +1,13 @@
+﻿using NUnit.Framework;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace HomesEnglandTest.UseCase.SaveUploadedAssetRegisterFile
+{
+    [TestFixture]
+    public class SaveUploadedAssetRegisterFileUseCaseTests
+    {
+
+    }
+}
