+++ WebApi/Controllers/AssetRegisterVersionController.cs
@@ -1,8 +1,12 @@
-﻿using System.Threading;
+﻿using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 using HomesEngland.UseCase.GetAsset.Models;
 using HomesEngland.UseCase.GetAssetRegisterVersions;
 using HomesEngland.UseCase.GetAssetRegisterVersions.Models;
+using HomesEngland.UseCase.SaveUploadedAssetRegisterFile;
+using HomesEnglandTest.UseCase.SaveUploadedAssetRegisterFile;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using WebApi.Extensions;
 
@@ -13,10 +17,12 @@ namespace WebApi.Controllers
     public class AssetRegisterVersionController : ControllerBase
     {
         private readonly IGetAssetRegisterVersionsUseCase _getAssetRegisterVersionsUseCase;
+        private readonly ISaveUploadedAssetRegisterFileUseCase _saveUploadedAssetRegisterFileUseCase;
 
-        public AssetRegisterVersionController(IGetAssetRegisterVersionsUseCase registerVersionsUseCase)
+        public AssetRegisterVersionController(IGetAssetRegisterVersionsUseCase registerVersionsUseCase, ISaveUploadedAssetRegisterFileUseCase saveUploadedAssetRegisterFileUseCase)
         {
             _getAssetRegisterVersionsUseCase = registerVersionsUseCase;
+            _saveUploadedAssetRegisterFileUseCase = saveUploadedAssetRegisterFileUseCase;
         }
 
         [HttpGet]
@@ -30,5 +36,8 @@ namespace WebApi.Controllers
             return this.StandardiseResponse<GetAssetRegisterVersionsResponse, AssetRegisterVersionOutputModel>(
                 await _getAssetRegisterVersionsUseCase.ExecuteAsync(request, CancellationToken.None).ConfigureAwait(false));
         }
+
+
+
     }
 }
