+++ WebApi/WebApi.csproj
@@ -25,6 +25,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\DependencyInjection\DependencyInjection.csproj" />
+    <ProjectReference Include="..\HomesEnglandTest\HomesEnglandTest.csproj" />
     <ProjectReference Include="..\HomesEngland\HomesEngland.csproj" />
     <ProjectReference Include="..\Infrastructure\Infrastructure.csproj" />
     <ProjectReference Include="..\Main\Main.csproj" />
