+++ WebApiTest/Controller/Asset/CalculateAssetAggregates/CalculateAssetAggregatesControllerTests.cs
@@ -5,9 +5,9 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using HomesEngland.Domain;
 using HomesEngland.Gateway.AssetRegisterVersions;
-using HomesEngland.UseCase.BulkCreateAsset.Models;
 using HomesEngland.UseCase.CalculateAssetAggregates;
 using HomesEngland.UseCase.CalculateAssetAggregates.Models;
+using HomesEngland.UseCase.CreateAssetRegisterVersion.Models;
 using HomesEngland.UseCase.SearchAsset.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
